+++ src/django_nh3/forms.py
@@ -1,4 +1,7 @@
-from typing import Any, Callable
+from __future__ import annotations
+
+from collections.abc import Callable
+from typing import Any
 
 import nh3
 from django import forms