+++ CHANGELOG.rst
@@ -4,3 +4,15 @@ Changelog
 
 unreleased
 ----------
+
+0.0.2 - 2023-11-15
+------------------
+
+- Added model field with NH3 sanitization
+- Added testing against django 5.0
+
+
+0.0.1 - 2023-03-02
+------------------
+
+- Initial release offering a form field with NH3 sanitization